Add new proyect struct
